passage aux requete HQL pour computer DAO suite

+++ computer-database-persistance/src/main/java/com/excilys/dao/IComputerDAO.java
@@ -36,7 +36,7 @@ public interface IComputerDAO {
 	 * @param filter
 	 * @return int
 	 */
-	public int count(String filter);
+	public long count(String filter);
 	
 	/**
 	 * Save a computer in the database
